align(core,event-store): AggregateVersion int→long to converge with 1.0.x
对齐决策：向 1.0.x 的宽类型收敛（long，事件溯源版本号无溢出风险）。
1.0.x 已是 long；2.0.x/3.0.x 同步升 long，消除三方共有符号中唯一
非允许轴的签名差异（AggregateVersion(int)→(long)）。

波及面 23 个文件统一升级：
- core AggregateVersion：字段 int→long，构造器参数 long
- core DomainEvent.getAggregateVersionInteger：返回 Integer→Long
- core AggregateVersionConflictException：构造参数 int→long
- 4 个 EventStore 实现（JPA/JDBI/R2DBC/Panache）+ ESDB + SPI（EventStore/AsyncEventStore）
  + InMemoryEventStore + EventStoreRetry：expectedVersion 参数 int→long
- 9 个测试类对齐

全量 mvn clean compile BUILD SUCCESS

+++ ddd4j-core/src/main/java/io/ddd4j/core/ddd/event/AggregateVersion.java
@@ -28,14 +28,14 @@ public final class AggregateVersion implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private final int value;
+    private final long value;
 
     /**
      * 创建版本号。
      *
      * @param value 大于等于零的版本号
      */
-    public AggregateVersion(int value) {
+    public AggregateVersion(long value) {
         if (value < 0) {
             throw new IllegalArgumentException("Aggregate version must not be negative");
         }
@@ -48,7 +48,7 @@ public final class AggregateVersion implements Serializable {
      * @return 版本号
      */
     @JsonValue
-    public int asInt() {
+    public long asInt() {
         return value;
     }
